Make splash a separate component file for now
